streaming/requirements.databricks.txt
pyspark==3.5.3
kafka-python==2.0.2
torch>=2.9,<2.10
torchvision>=0.24,<0.25
Pillow>=11.2,<13
